チャートレイアウトとスキル登録のカスタムフック作成
